refactor: replace custom error extraction functions with centralized error handling in Discord and Slack integrations
