+++ integrations/discord/delivery.py
@@ -3,10 +3,10 @@
 from __future__ import annotations
 
 import logging
-from collections.abc import Mapping
 from typing import Any
 
 from platform.common.truncation import truncate
+from platform.notifications.delivery_errors import extract_http_error
 from platform.notifications.delivery_transport import post_json
 
 logger = logging.getLogger(__name__)
@@ -23,24 +23,6 @@ def _redact_token(text: str, bot_token: str) -> str:
     if bot_token and bot_token in text:
         return text.replace(bot_token, "<redacted>")
     return text
-
-
-def _extract_error(data: Mapping[str, Any], status_code: int, text: str) -> str:
-    """Return a human-readable error string from a Discord API response.
-
-    Tries ``data["message"]`` / ``data["error"]`` first, then falls back to
-    the raw response body or the HTTP status code so non-JSON failure bodies
-    (HTML, plain text) never cause a crash.
-    """
-    msg = data.get("message")
-    if msg:
-        return str(msg)
-    err = data.get("error")
-    if err:
-        return str(err)
-    if text:
-        return text[:500]
-    return f"HTTP {status_code}"
 
 
 def post_discord_message(
@@ -65,7 +47,7 @@ def post_discord_message(
         return False, safe_error, ""
     if response.status_code not in (200, 201):
         logger.warning("[discord] post message failed: %s", response.status_code)
-        error_message = _extract_error(response.data, response.status_code, response.text)
+        error_message = extract_http_error(response.data, response.status_code, response.text)
         safe_error = _redact_token(error_message, bot_token)
         logger.warning("[discord] post message failed: %s", safe_error)
         return False, safe_error, ""
@@ -93,7 +75,7 @@ def create_discord_thread(
         logger.warning("[discord] create thread exception: %s", safe_error)
         return False, safe_error, ""
     if response.status_code not in (200, 201):
-        error_message = _extract_error(response.data, response.status_code, response.text)
+        error_message = extract_http_error(response.data, response.status_code, response.text)
         safe_error = _redact_token(error_message, bot_token)
         logger.warning("[discord] create thread failed: %s", safe_error)
         return False, safe_error, ""
