+++ integrations/slack/delivery.py
@@ -8,6 +8,7 @@ import re
 from typing import Any
 
 from config.config import SLACK_CHANNEL
+from platform.notifications.delivery_errors import extract_http_error
 from platform.notifications.delivery_transport import post_json
 from platform.observability import debug_print
 
@@ -22,21 +23,6 @@ def _redact_token(text: str, access_token: str) -> str:
     if access_token and access_token in text:
         redacted = text.replace(access_token, "<redacted>")
     return _ACCESS_TOKEN_RE.sub(r"\1<redacted>", redacted)
-
-
-def _extract_error(data: dict[str, Any], status_code: int, text: str) -> str:
-    """Return a human-readable error string from a Slack API response.
-
-    Tries ``data["error"]`` first, then falls back to the raw response body
-    or the HTTP status code so non-JSON failure bodies (HTML, plain text)
-    never cause a crash.
-    """
-    error = data.get("error")
-    if error:
-        return str(error)
-    if text:
-        return text[:500]
-    return f"HTTP {status_code}"
 
 
 def _slack_bearer_headers(token: str) -> dict[str, str]:
@@ -313,7 +299,7 @@ def _post_direct(
     if response.data.get("ok") is not True:
         error = response.data.get("error")
         if not error:
-            error = _extract_error(dict(response.data), response.status_code, response.text)
+            error = extract_http_error(response.data, response.status_code, response.text)
         safe_error = _redact_token(str(error), token)
         response_meta = response.data.get("response_metadata", {})
         logger.error(
